requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
requests==2.31.0
python-dotenv==1.0.1
Flask-Mail==0.9.1
werkzeug==3.1.3
newspaper3k==0.2.8
python-dateutil==2.8.2 
gunicorn==21.2.0
Flask-Caching==2.1.0
Flask-Compress==1.14
beautifulsoup4==4.12.3
